requirements.txt
aniso8601==6.0.0
Click==7.0
Flask==1.0.2
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
get==2019.4.13
itsdangerous==1.1.0
Jinja2==2.10.1
jsonify==0.5
MarkupSafe==1.1.1
mysql-connector-python==8.0.16
post==2019.4.13
protobuf==3.7.1
public==2019.4.13
pytz==2019.1
query-string==2019.4.13
request==2019.4.13
six==1.12.0
SQLAlchemy==1.3.3
Werkzeug==0.15.2
